Adding Chevrolet Agile, Chevrolet Celta, Chevrolet Cobalt, and Chevrolet Corvette

+++ catscripts/Entertainment/Puzzles/Puzzles.sh
@@ -1,26 +1,36 @@
 #!/bin/bash
 
-egrep -i 'Sudoku' newpages.txt >> Sudoku.txt
-egrep -i 'Crossword' newpages.txt >> Crossword.txt
-egrep -i '[0-9]{1,}( |-)letter word for' newpages.txt >> Crossword.txt
-egrep -i '(three|four|five|six|seven|eight|nine|ten|eleven|twelve|thirteen|fourteen|fifteen|sixteen|seventeen|eighteen|nineteen|twenty)( |-)letter word for' newpages.txt >> Crossword.txt
+KEYWORDS_SUDOKU="Sudoku"
+KEYWORDS_CROSSWORD="Crossword|[0-9]{1,}(| |-)letter(| )word(| )for|'(three|four|five|six|seven|eight|nine|ten|eleven|twelve|thirteen|fourteen|fifteen|sixteen|seventeen|eighteen|nineteen|twenty)( |-)letter word for"
 
-SUDOKU=`stat --print=%s Sudoku.txt`
-CROSSWORD=`stat --print=%s Crossword.txt`
-
-if [ $SUDOKU -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Sudoku.txt"
-  export CATNAME="Sudoku"
-  $CATEGORIZE
-fi
+  
+  debug_start "Puzzles"
 
-if [ $CROSSWORD -ne 0 ];
-then
-  export CATFILE="Crossword.txt"
-  export CATNAME="Crossword puzzles"
-  $CATEGORIZE
-fi
+  SUDOKU="$(egrep -i "$KEYWORDS_SUDOKU" newpages.txt)"
+  CROSSWORD="$(egrep -i "$KEYWORDS_CROSSWORD" newpages.txt)"
 
-rm Sudoku.txt
-rm Crossword.txt
+  if [ "$SUDOKU" != "" ];
+  then
+    printf "%s" "$SUDOKU" > Sudoku.txt
+    export CATFILE="Sudoku.txt"
+    export CATNAME="Sudoku"
+    $CATEGORIZE
+    rm Sudoku.txt
+    unset SUDOKU
+  fi
+
+  if [ "$CROSSWORD" != "" ];
+  then
+    printf "%s" "$CROSSWORD" > Crossword.txt
+    export CATFILE="Crossword.txt"
+    export CATNAME="Crossword puzzles"
+    $CATEGORIZE
+    rm Crossword.txt
+    unset CROSSWORD
+  fi
+
+  debug_end "Puzzles"
+
+fi
