+++ catscripts/Government/Countries/United_States/US_states/New_York/NewYork.sh
@@ -1,24 +1,37 @@
 #!/bin/bash
 
-egrep -i 'New(| )York' newpages.txt | egrep -iv 'New(| )York(| )City|\bNYC|Manhattan|Brooklyn' >> NewYork.txt
-egrep -i 'New(| )York(| )City|\bNYC|Manhattan' newpages.txt >> NewYorkCity.txt
+KEYWORDS_NEWYORK="New(| )York"
+KEYWORDS_NYC="New(| )York(| )City|\bNYC|New(| )York(|,)(| )New(| )York|Manhattan|Brooklyn|Queen(|,)(| )New(| )York"
+KEYWORDS_NEWYORK_EXCLUDE="$KEYWORDS_NYC"
 
-NEWYORK=`stat --print=%s NewYork.txt`
-NYC=`stat --print=%s NewYorkCity.txt`
-
-if [ $NEWYORK -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="NewYork.txt"
-  export CATNAME="New York"
-  $CATEGORIZE
-fi
 
-if [ $NYC -ne 0 ];
-then
-  export CATFILE="NewYorkCity.txt"
-  export CATNAME="New York City"
-  $CATEGORIZE
-fi
+  debug_start "New York"
 
-rm NewYork.txt
-rm NewYorkCity.txt
+  NEWYORK="$(egrep -i "$KEYWORDS_NEWYORK" newpages.txt | egrep -iv "$KEYWORDS_NEWYORK_EXCLUDE")"
+  NYC="$(egrep -i "$KEYWORDS_NYC" newpages.txt)"
+
+  if [ "$NEWYORK" != "" ];
+  then
+    printf "%s" "$NEWYORK" > NewYork.txt
+    export CATFILE="NewYork.txt"
+    export CATNAME="New York"
+    $CATEGORIZE
+    rm NewYork.txt
+    unset NEWYORK
+  fi
+
+  if [ "$NYC" != "" ];
+  then
+    printf "%s" "$NYC" > NewYorkCity.txt
+    export CATFILE="NewYorkCity.txt"
+    export CATNAME="New York City"
+    $CATEGORIZE
+    rm NewYorkCity.txt
+    unset NYC
+  fi
+
+  debug_end "New York"
+
+fi
